pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
pymongo = "^4.0"
discord = "^1.7.3"
"dislash.py" = "^1.4.9"
dnspython = "^2.1.0"
Flask = "^2.0.2"
openai = "^0.26.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
